[css-vars-theme] Convert border and sizing to CSS variables
