+++ __docs__/wonder-blocks-tokens/tokens-border.mdx
@@ -4,7 +4,7 @@ import TokenTable from "../components/token-table";
 
 import Banner from "@khanacademy/wonder-blocks-banner";
 import {View} from "@khanacademy/wonder-blocks-core";
-import * as tokens from "@khanacademy/wonder-blocks-tokens";
+import {border, color, sizing} from "@khanacademy/wonder-blocks-tokens";
 
 <Meta title="Packages / Tokens / Border" />
 
@@ -18,6 +18,10 @@ export const baseColumns = (kind) => [
         ),
     },
     {
+        label: "CSS Variable",
+        cell: (row) => <code>{row.css}</code>,
+    },
+    {
         label: "Value",
         cell: "value",
     },
@@ -28,7 +32,7 @@ export const baseColumns = (kind) => [
 The `border` tokens are used to define the border properties of an element. The
 border tokens are broken down into two categories: `radius` and `width`.
 
-<View style={{marginBottom: tokens.spacing.medium_16}}>
+<View style={{marginBottom: sizing.size_160}}>
     <Banner
         kind="info"
         text="We don't define border colors here, and instead we recommend you to rely on the `color` token primitives to build on top of that."
@@ -42,8 +46,8 @@ border tokens are broken down into two categories: `radius` and `width`.
 import {border} from "@khanacademy/wonder-blocks-tokens";
 const styles = {
     borderedContainer: {
-        borderRadius: border.radius.small,
-        borderWidth: border.width.small,
+        borderRadius: border.radius.radius_040,
+        borderWidth: border.width.thin,
     },
 };
 ```
@@ -60,10 +64,10 @@ const styles = {
             cell: (row) => (
                 <View
                     style={{
-                        backgroundColor: tokens.color.purple,
+                        backgroundColor: color.purple,
                         borderRadius: row.value,
-                        width: tokens.sizing.size_480,
-                        height: tokens.sizing.size_480,
+                        width: sizing.size_480,
+                        height: sizing.size_480,
                     }}
                 >
                     &nbsp;
@@ -71,7 +75,7 @@ const styles = {
             ),
         },
     ]}
-    tokens={tokens.border.radius}
+    tokens={border.radius}
 />
 
 ### Width
@@ -84,10 +88,10 @@ const styles = {
             cell: (row) => (
                 <View
                     style={{
-                        borderColor: tokens.color.purple,
+                        borderColor: color.purple,
                         borderWidth: row.value,
-                        width: tokens.sizing.size_480,
-                        height: tokens.sizing.size_480,
+                        width: sizing.size_480,
+                        height: sizing.size_480,
                     }}
                 >
                     &nbsp;
@@ -95,5 +99,5 @@ const styles = {
             ),
         },
     ]}
-    tokens={tokens.border.width}
+    tokens={border.width}
 />
