+++ packages/wonder-blocks-tokens/src/theme/default.ts
@@ -1,0 +1,9 @@
+import {sizing} from "./primitive/sizing";
+import {semanticColor} from "./semantic/semantic-color";
+import {border} from "./primitive/border";
+
+export default {
+    border,
+    semanticColor,
+    sizing,
+};
