+++ packages/wonder-blocks-tokens/src/theme/primitive/border.ts
@@ -1,4 +1,5 @@
-import {sizing, remToPx} from "./sizing";
+import {remToPx} from "../../util/sizing-utils";
+import {sizing} from "./sizing";
 
 export const border = {
     /**
