--- packages/wonder-blocks-tokens/src/tokens/semantic-color.ts
@@ -1,5 +1,0 @@
-import {CSS_VAR_COLOR_PREFIX} from "../util/constants";
-import {mapValuesToCssVars} from "../internal/map-values-to-css-vars";
-import tokens from "../theme/color/default";
-
-export const semanticColor = mapValuesToCssVars(tokens, CSS_VAR_COLOR_PREFIX);
